+++ code/src/sixsq/nuvla/auth/oidc.clj
@@ -1,12 +1,12 @@
 (ns sixsq.nuvla.auth.oidc
   (:require
+    [buddy.sign.jws :as jws]
     [clj-http.client :as http]
     [clojure.data.json :as json]
     [clojure.pprint :refer [pprint]]
     [clojure.tools.logging :as log]))
 
-
-(defn get-access-token
+(defn get-id-token
   "Perform an HTTP POST to the OIDC/MitreID server to recover an access token.
    This method will log exceptions but then return nil to indicate that the
    access token could not be retrieved."
@@ -21,12 +21,38 @@
                                   :client_secret client-secret}})
         :body
         (json/read-str :key-fn keyword)
-        :access_token)
+        :id_token)
     (catch Exception e
       (let [client-secret? (str (boolean client-secret))]
         (if-let [{:keys [status] :as data} (ex-data e)]
           (log/errorf "error status %s getting access token from %s with client_id %s, code %s, and client_secret %s\n%s"
                       status tokenURL client-id oidc-code client-secret? (with-out-str (pprint data)))
           (log/errorf "unexpected error when getting access token from %s with client_id %s, code %s, and client_secret %s\n%s"
-                      tokenURL client-id oidc-code client-secret? (str e))))
-      nil)))
+                      tokenURL client-id oidc-code client-secret? (str e)))))))
+
+(defn get-kid-from-id-token
+  [id-token]
+  (try
+    (-> id-token jws/decode-header :kid)
+    (catch Exception e
+      (log/errorf "OIDC extraction of kid from id-token failed: %s (%s)"
+                  id-token (str e)))))
+
+(defn get-public-key
+  [jwks-url kid]
+  (log/debugf "getting public key from jwks-url='%s' with kid='%s'"
+              jwks-url kid)
+  (try
+    (-> (http/get jwks-url {:headers {"Accept" "application/json"}})
+        :body
+        (json/read-str :key-fn keyword)
+        :keys
+        (->> (some #(when (= (:kid %) kid) (json/write-str %)))))
+    (catch Exception e
+      (if-let [{:keys [status] :as data} (ex-data e)]
+        (log/errorf
+          "OIDC jwks public key failed with status %s from %s \n%s"
+          status jwks-url (with-out-str (pprint data)))
+        (log/errorf
+          "OIDC unexpected error when getting public key from %s \n%s"
+          jwks-url (str e))))))