+++ code/src/sixsq/nuvla/server/resources/callback_create_session_mitreid.clj
@@ -28,7 +28,7 @@
         {:keys [instance]} (crud/retrieve-by-id-as-admin href)
         {:keys [client-id client-secret public-key token-url]} (oidc-utils/config-mitreid-params redirect-url instance)]
     (if-let [code (uh/param-value request :code)]
-      (if-let [access-token (auth-oidc/get-access-token client-id client-secret token-url code (str base-uri (or callback-id "unknown-id") "/execute"))]
+      (if-let [access-token (auth-oidc/get-id-token client-id client-secret token-url code (str base-uri (or callback-id "unknown-id") "/execute"))]
         (try
           (let [{:keys [sub] :as claims} (sign/unsign-cookie-info access-token public-key)
                 roles (concat (oidc-utils/extract-roles claims)