+++ code/src/sixsq/nuvla/server/resources/hook_oidc_session.clj
@@ -30,14 +30,14 @@ Stripe oidc session.
         instance          (get params :instance oidc-utils/geant-instance)
         redirect-hook-url (cond-> (str base-uri "hook" "/" action)
                                   (not= instance oidc-utils/geant-instance) (str "/" instance))
-        {:keys [client-id client-secret
-                public-key token-url]} (oidc-utils/config-oidc-params redirect-ui-url instance)]
+        {:keys [client-id client-secret token-url jwks-url]} (oidc-utils/config-oidc-params redirect-ui-url instance)]
     (log/info "hook-oidc-session redirect request:" request)
     (if-let [code (uh/param-value request :code)]
-      (if-let [access-token (auth-oidc/get-access-token client-id client-secret token-url
-                                                        code redirect-hook-url)]
+      (if-let [id-token (auth-oidc/get-id-token client-id client-secret token-url
+                                                code redirect-hook-url)]
         (try
-          (let [{:keys [sub] :as claims} (sign/unsign-cookie-info access-token public-key)
+          (let [public-key (auth-oidc/get-public-key id-token jwks-url)
+                {:keys [sub] :as claims} (sign/unsign-cookie-info id-token public-key)
                 roles (concat (oidc-utils/extract-roles claims)
                               (oidc-utils/extract-groups claims)
                               (oidc-utils/extract-entitlements claims))]