+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+### Changed
+
+- Nuvlabox status - next-heartbeat should not be changed when online is set to
+  false by admin
+
 ## [6.0.2] - 2022-08-03
 
 ### Added